Update url container to match spec
