+++ apps/kg-explorer/src/app/pages/metadata-page/metadata-page.component.ts
@@ -8,7 +8,7 @@ import { DigitalObjectsJsonLd, V1Service } from '@hra-api/ng-client';
 import { watchBreakpoint } from '@hra-ui/cdk/breakpoints';
 import { PageSectionComponent } from '@hra-ui/design-system/content-templates/page-section';
 import { FooterComponent } from '@hra-ui/design-system/navigation/footer';
-import { PurlContainerComponent } from '@hra-ui/design-system/purl-container';
+import { CopyableUrlContainerComponent } from '@hra-ui/design-system/copyable-url-container';
 import { MenuOptionsType, TableColumn } from '@hra-ui/design-system/table';
 import { MarkdownComponent } from 'ngx-markdown';
 
@@ -30,7 +30,7 @@ import { getOrganIcon, getProductIcon, getProductLabel, sentenceCase } from '../
     ProvenanceMenuComponent,
     MatChipsModule,
     FooterComponent,
-    PurlContainerComponent,
+    CopyableUrlContainerComponent,
   ],
   templateUrl: './metadata-page.component.html',
   styleUrl: './metadata-page.component.scss',
