+++ libs/design-system/copyable-url-container/ng-package.json
