+++ libs/design-system/copyable-url-container/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/copyable-url-container.component';
