--- libs/design-system/purl-container/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/purl-container.component';
