--- libs/design-system/purl-container/src/lib/purl-container.component.html
@@ -1,18 +1,0 @@
-<a
-  class="purl-link"
-  mat-button
-  hraButtonVariant="secondary"
-  [attr.href]="purl()"
-  target="_blank"
-  rel="noopener noreferrer"
-  hraPlainTooltip="Open in new tab"
->
-  <mat-icon>open_in_new</mat-icon>
-  <span class="purl-text">
-    {{ purl() }}
-  </span>
-</a>
-<button class="purl-copy" mat-button hraPlainTooltip="Copy link" (click)="copyPurl()">
-  <mat-icon>content_copy</mat-icon>
-  Copy
-</button>
